Addressed comments on last revision request
