+++ pyproject.toml
@@ -18,12 +18,15 @@ license = {text = "MIT"}
 
 dynamic = ["version", "readme", "dependencies", "optional-dependencies"]
 
+[metadata]
+description = "Flexible, portable, and efficient geospatial evaluations for a variety of data."
+
 
 [tool.setuptools.dynamic]
 dependencies = {file = ["./requirements/base.txt"]}
 #optional-dependencies = {file = ["./requirements/test.txt"]}
 version = {file = ["./gval/version.txt"]}
-readme = {file = ["README.MD"]}
+readme = {file = ["README.MD"], content-type = "text/markdown"}
 
 
 #[tool.setuptools.packages.find]
