+++ requirements/base.txt
@@ -2,6 +2,7 @@ rioxarray==0.14.0
 dask==2023.1.1
 xarray-spatial==0.3.5
 pandera==0.14.5
+shapely==2.0.1
 
 ###################################
 # Potential future packages
